GifTastic
# gifTastic
This project uses API's to populate the screen with gifs whenever the user clicks on a button. 
The user can also create a button and submit it as well. 
The data still and data animate attributes were added in order for each gif to animate and pause on the user click as well.